Introduced extension aliases.
When several filename extensions corresponded to the same programming language,
the compiler and runtime settings had to be duplicated.

It is now possible to define one extension as an alias for another, so that
it is used only for locating the source. The copy of the program in $TDIR
already has the new extension and it is processed according to it.